backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
email-validator
httpx
supabase==2.10.0
gotrue==2.10.0
google-generativeai
openpyxl
pdfplumber
python-multipart
PyJWT
pytest
